python-api/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
PyMuPDF
Pillow
PyPDF2
reportlab
